Facilitate the use of constraints.
Attempting to use constraints in avatars was broken due to how we copy
the pose of the built-in model onto our provided custom models in
`LateUpdate`, _after_ it seems the
`PreLateUpdate.ConstraintManagerUpdate` has fired.

So, what we do here: Introduce another event which fires before the
`[...].ConstraintManagerUpdate`, and have the pose copy performed under
this new event, such that constraints can operate on the
intended/desired pose.